Fixing some benign type mismatches and removing unused variables

+++ src/derived_types.f90
@@ -50,7 +50,7 @@ module derived_types
     ! Concentrations of the chemical species
     real(real64), dimension(:), allocatable :: species_concentrations
     ! Number of atoms of each chemical species
-    integer(real64), dimension(:), allocatable :: species_numbers
+    integer(int64), dimension(:), allocatable :: species_numbers
     ! Atom-atom interchange interaction file name
     character(len=50) :: interaction_file
     ! Interaction range (number of coordination shells)
